test: Move some helper functions from `core` to `keycloak` test fixtures
This makes them available to the `plugin-manager` component.
