+++ clients/keycloak/src/testFixtures/kotlin/Extensions.kt
@@ -17,6 +17,8 @@
  * License-Filename: LICENSE
  */
 
+@file:Suppress("TooManyFunctions")
+
 package org.eclipse.apoapsis.ortserver.clients.keycloak.test
 
 import dasniko.testcontainers.keycloak.KeycloakContainer
@@ -31,8 +33,11 @@ import org.eclipse.apoapsis.ortserver.clients.keycloak.Role
 import org.eclipse.apoapsis.ortserver.clients.keycloak.RoleName
 import org.eclipse.apoapsis.ortserver.clients.keycloak.User
 
+import org.keycloak.admin.client.Keycloak
+import org.keycloak.representations.idm.ClientScopeRepresentation
 import org.keycloak.representations.idm.CredentialRepresentation
 import org.keycloak.representations.idm.GroupRepresentation
+import org.keycloak.representations.idm.ProtocolMapperRepresentation
 import org.keycloak.representations.idm.RoleRepresentation
 import org.keycloak.representations.idm.UserRepresentation
 
@@ -137,3 +142,92 @@ fun KeycloakContainer.createKeycloakConfigMapForTestRealm() =
  */
 fun KeycloakContainer.createKeycloakClientForTestRealm() =
     DefaultKeycloakClient.create(createKeycloakClientConfigurationForTestRealm(), Json { ignoreUnknownKeys = true })
+
+/**
+ * Create a map containing JWT configuration properties for the [testRealm] and this [KeycloakContainer]. The map can be
+ * used to configure a Ktor test application.
+ */
+fun KeycloakContainer.createJwtConfigMapForTestRealm() =
+    mapOf(
+        "jwt.jwksUri" to "$authServerUrl/realms/$TEST_REALM/protocol/openid-connect/certs",
+        "jwt.issuer" to "$authServerUrl/realms/$TEST_REALM",
+        "jwt.realm" to TEST_REALM,
+        "jwt.audience" to TEST_SUBJECT_CLIENT,
+        "jwt.roleCacheLifetimeSeconds" to "0"
+    )
+
+/**
+ * Add the provided [role] in the [TEST_SUBJECT_CLIENT] to the provided [username].
+ */
+fun Keycloak.addUserRole(username: String, role: String) {
+    realm(TEST_REALM).apply {
+        val client = clients().findByClientId(TEST_SUBJECT_CLIENT).single()
+        val user = users().search(username).single()
+
+        val roleRepresentation = clients().get(client.id).roles().get(role).toRepresentation()
+        users().get(user.id).roles().clientLevel(client.id).add(listOf(roleRepresentation))
+    }
+}
+
+/**
+ * Create an audience mapper that adds the provided [audience] to the access token.
+ */
+private fun audienceMapper(audience: String) = ProtocolMapperRepresentation().apply {
+    name = "audience-mapper"
+    protocol = "openid-connect"
+    protocolMapper = "oidc-audience-mapper"
+    config = mapOf(
+        "included.client.audience" to audience,
+        "id.token.claim" to "false",
+        "access.token.claim" to "true",
+        "userinfo.token.claim" to "false"
+    )
+}
+
+/**
+ * Create a client scope that maps the provided [audience] to the audience in the JWT token and add the created scope to
+ * the default scopes of the [TEST_CLIENT].
+ */
+fun Keycloak.setUpClientScope(audience: String) {
+    val subjectClientScope = "$TEST_SUBJECT_CLIENT-scope"
+
+    realm(TEST_REALM).apply {
+        clientScopes().create(
+            ClientScopeRepresentation().apply {
+                name = subjectClientScope
+                protocol = "openid-connect"
+                protocolMappers = listOf(audienceMapper(audience))
+            }
+        )
+
+        val testClient = clients().get(clients().findByClientId(TEST_CLIENT).first().id)
+        val clientScope = clientScopes().findAll().single { it.name == subjectClientScope }
+        testClient.addDefaultClientScope(clientScope.id)
+    }
+}
+
+/**
+ * Create a [user] with the provided [password].
+ */
+fun Keycloak.setUpUser(user: User, password: String) {
+    realm(TEST_REALM).users().create(user.toUserRepresentation(password = password))
+}
+
+/**
+ * Create the provided [roles] in the [TEST_SUBJECT_CLIENT] and assign them to the provided [username].
+ */
+fun Keycloak.setUpUserRoles(username: String, roles: List<String>) {
+    realm(TEST_REALM).apply {
+        val client = clients().findByClientId(TEST_SUBJECT_CLIENT).single()
+
+        val roleRepresentations = roles.map { role ->
+            clients().get(client.id).roles().run {
+                create(RoleRepresentation().apply { name = role })
+                get(role).toRepresentation()
+            }
+        }
+
+        val user = users().search(username).single()
+        users().get(user.id).roles().clientLevel(client.id).add(roleRepresentations)
+    }
+}
